backend/requirements.txt
fastapi==0.68.1
python-multipart==0.0.5
boto3==1.26.137
PyMuPDF==1.21.1
pymupdf==1.21.1
python-dotenv==0.19.2
langchain==0.0.184
faiss-cpu==1.7.3
openai==0.27.4
uvicorn==0.15.0